No need for NgComponentOutlet
Instead of having two layers of lazy-loading, one for the component
`DynamicComponent` via NgComponentOutlet, and one for Highcharts itself
via `DynamicComponent.ngOnInit()`, we can have only one for Highcharts.

This is easier to use, since there is nothing special to do when using
`DynamicComponent`, and we still benefit from lazy-loading Highcharts.

And the `@if` is not necessary to have lazy-loading in both dev build
and production build. So it can be removed entirely.